CR-869 RHUI Unlinked UAC Incomplete - Made confirm-address common
