+++ app/requests_handlers.py
@@ -120,125 +120,6 @@ class RequestCode(RequestCommon):
 
 
 @requests_routes.view(r'/' + View.valid_display_regions + '/requests/' +
-                      RequestCommon.valid_request_types + '-code/confirm-address/')
-class RequestCodeConfirmAddress(RequestCommon):
-    @aiohttp_jinja2.template('request-code-confirm-address.html')
-    async def get(self, request):
-        self.setup_request(request)
-
-        request_type = request.match_info['request_type']
-        display_region = request.match_info['display_region']
-
-        if display_region == 'cy':
-            page_title = "Ydy'r cyfeiriad hwn yn gywir?"
-            locale = 'cy'
-        else:
-            page_title = 'Is this address correct?'
-            locale = 'en'
-
-        self.log_entry(request, display_region + '/requests/' + request_type + '-code/confirm-address')
-
-        attributes = await self.get_check_attributes(request, request_type, display_region)
-        attributes['page_title'] = page_title
-        attributes['display_region'] = display_region
-        attributes['locale'] = locale
-        attributes['request_type'] = request_type
-
-        return attributes
-
-    @aiohttp_jinja2.template('request-code-confirm-address.html')
-    async def post(self, request):
-        self.setup_request(request)
-
-        request_type = request.match_info['request_type']
-        display_region = request.match_info['display_region']
-
-        if display_region == 'cy':
-            page_title = "Ydy'r cyfeiriad hwn yn gywir?"
-            locale = 'cy'
-        else:
-            page_title = 'Is this address correct?'
-            locale = 'en'
-
-        self.log_entry(request, display_region + '/requests/' + request_type + '-code/confirm-address')
-
-        attributes = await self.get_check_attributes(request, request_type, display_region)
-
-        attributes['page_title'] = page_title
-        attributes['display_region'] = display_region
-        attributes['locale'] = locale
-        attributes['request_type'] = request_type
-
-        data = await request.post()
-
-        try:
-            address_confirmation = data['form-confirm-address']
-        except KeyError:
-            logger.info('address confirmation error',
-                        client_ip=request['client_ip'])
-            if display_region == 'cy':
-                flash(request, ADDRESS_CHECK_MSG_CY)
-            else:
-                flash(request, ADDRESS_CHECK_MSG)
-            return attributes
-
-        if address_confirmation == 'yes':
-
-            session = await get_session(request)
-            uprn = session['attributes']['uprn']
-
-            uprn_ai_return = await AddressIndex.get_ai_uprn(request, uprn)
-
-            try:
-                if uprn_ai_return['response']['address']['countryCode'] == 'S':
-                    logger.info('address is in Scotland', client_ip=request['client_ip'])
-                    raise HTTPFound(
-                        request.app.router['CommonAddressInScotland:get'].
-                        url_for(display_region=display_region, user_journey='requests', request_type=request_type))
-            except KeyError:
-                logger.info('unable to check for region', client_ip=request['client_ip'])
-
-            # uprn_return[0] will need updating/changing for multiple households - post 2019 issue
-            try:
-                uprn_return = await RHService.get_cases_by_uprn(request, uprn)
-                session['attributes']['case_id'] = uprn_return[0]['caseId']
-                session['attributes']['region'] = uprn_return[0]['region']
-                session.changed()
-                raise HTTPFound(
-                    request.app.router['RequestCodeEnterMobile:get'].
-                    url_for(request_type=request_type, display_region=display_region))
-            except ClientResponseError as ex:
-                if ex.status == 404:
-                    logger.info('unable to match uprn',
-                                client_ip=request['client_ip'])
-                    raise HTTPFound(
-                        request.app.router['CommonCallContactCentre:get'].
-                        url_for(request_type=request_type,
-                                user_journey='requests',
-                                display_region=display_region,
-                                error='unable-to-match-address'))
-                else:
-                    raise ex
-
-        elif address_confirmation == 'no':
-            raise HTTPFound(
-                request.app.router['CommonEnterAddress:get'].url_for(sub_user_journey=request_type + '-code',
-                                                                     display_region=display_region,
-                                                                     user_journey='requests'))
-
-        else:
-            # catch all just in case, should never get here
-            logger.info('address confirmation error',
-                        client_ip=request['client_ip'])
-            flash(request, ADDRESS_CHECK_MSG)
-            attributes['page_title'] = page_title
-            attributes['display_region'] = display_region
-            attributes['locale'] = locale
-            attributes['request_type'] = request_type
-            return attributes
-
-
-@requests_routes.view(r'/' + View.valid_display_regions + '/requests/' +
                       RequestCommon.valid_request_types + '-code/enter-mobile/')
 class RequestCodeEnterMobile(RequestCommon):
     @aiohttp_jinja2.template('request-code-enter-mobile.html')
