+++ app/templates/common-confirm-address.html
@@ -1,0 +1,80 @@
+{% extends 'base-' + display_region + '.html' %}
+
+{% set messages_dict=dict(get_flashed_messages()|groupby('level')) %}
+{% set field_messages_dict=dict(messages_dict['ERROR']|groupby('field')) %}
+
+{% from 'components/question/_macro.njk' import onsQuestion %}
+{% from 'components/button/_macro.njk' import onsButton %}
+{% from 'components/radios/_macro.njk' import onsRadios %}
+
+{% set form =  {
+    'method': 'POST',
+    'attributes': {
+        'action': url('CommonConfirmAddress:post', display_region=display_region, user_journey=user_journey, sub_user_journey=sub_user_journey)
+    }
+} %}
+
+{% set form_options = [
+                {
+                    'id': 'yes',
+                    'label': {
+                        'text': _('Yes, this address is correct')
+                    },
+                    'value': 'yes'
+                },
+                {
+                    'id': 'no',
+                    'label': {
+                        'text': _('No, I need to change it')
+                    },
+                    'value': 'no'
+                }
+            ]
+%}
+
+{% block main %}
+
+        {% if messages_dict %}
+        {% include 'partials/messages.html' with context %}
+    {% endif %}
+
+    {% call onsQuestion({
+        'title': _('Is this address correct?')
+    }) %}
+
+    <p>
+        {{ addressLine1 }}<br>
+        {% if addressLine2 !='' %}{{ addressLine2 }}<br>{% endif %}
+        {% if addressLine3 !='' %}{{ addressLine3 }}<br>{% endif %}
+        {% if townName !='' %}{{ townName }}<br>{% endif %}
+        {{ postcode }}
+    </p>
+
+{% if 'address' in field_messages_dict %}
+    {{
+        onsRadios({
+            'name': 'form-confirm-address',
+            'radios': form_options,
+            'error': {
+                'text': _('Check and confirm the address')
+            }
+        })
+    }}
+{% else %}
+    {{
+        onsRadios({
+            'name': 'form-confirm-address',
+            'radios': form_options
+        })
+    }}
+{% endif %}
+    {% endcall %}
+
+    {{
+        onsButton({
+            'text': _('Continue'),
+            'classes': 'u-mt-l'
+        })
+    }}
+
+{% endblock %}
