--- app/templates/partials/confirm-address.html
@@ -1,42 +1,0 @@
-    {% if messages_dict %}
-        {% include 'partials/messages.html' with context %}
-    {% endif %}
-
-    {% call onsQuestion({
-        'title': _('Is this address correct?')
-    }) %}
-
-    <p>
-        {{ addressLine1 }}<br>
-        {% if addressLine2 !='' %}{{ addressLine2 }}<br>{% endif %}
-        {% if addressLine3 !='' %}{{ addressLine3 }}<br>{% endif %}
-        {% if townName !='' %}{{ townName }}<br>{% endif %}
-        {{ postcode }}
-    </p>
-
-{% if 'address' in field_messages_dict %}
-    {{
-        onsRadios({
-            'name': 'form-confirm-address',
-            'radios': form_options,
-            'error': {
-                'text': _('Check and confirm the address')
-            }
-        })
-    }}
-{% else %}
-    {{
-        onsRadios({
-            'name': 'form-confirm-address',
-            'radios': form_options
-        })
-    }}
-{% endif %}
-    {% endcall %}
-
-    {{
-        onsButton({
-            'text': _('Continue'),
-            'classes': 'u-mt-l'
-        })
-    }}
