--- app/templates/request-code-confirm-address.html
@@ -1,39 +1,0 @@
-{% extends 'base-' + display_region + '.html' %}
-
-{% set messages_dict=dict(get_flashed_messages()|groupby('level')) %}
-{% set field_messages_dict=dict(messages_dict['ERROR']|groupby('field')) %}
-
-{% from 'components/question/_macro.njk' import onsQuestion %}
-{% from 'components/button/_macro.njk' import onsButton %}
-{% from 'components/radios/_macro.njk' import onsRadios %}
-
-{% set form =  {
-    'method': 'POST',
-    'attributes': {
-        'action': url('RequestCodeConfirmAddress:post', request_type=request_type, display_region=display_region)
-    }
-} %}
-
-{% set form_options = [
-                {
-                    'id': 'yes',
-                    'label': {
-                        'text': _('Yes, this address is correct')
-                    },
-                    'value': 'yes'
-                },
-                {
-                    'id': 'no',
-                    'label': {
-                        'text': _('No, I need to change it')
-                    },
-                    'value': 'no'
-                }
-            ]
-%}
-
-{% block main %}
-
-    {% include 'partials/confirm-address.html' with context %}
-
-{% endblock %}
