+++ tests/unit/__init__.py
@@ -656,23 +656,15 @@ class RHTestCase(AioHTTPTestCase):
         self.post_request_code_select_address_hh_ni = self.app.router['CommonSelectAddress:post'].url_for(
             display_region='ni', user_journey='requests', sub_user_journey='household-code'
         )
-        self.get_requestcode_address_confirmation_hh_en = self.app.router['RequestCodeConfirmAddress:get'].url_for(
-            request_type='household', display_region='en'
+
+        self.post_request_code_confirm_address_hh_en = self.app.router['CommonConfirmAddress:post'].url_for(
+            display_region='en', user_journey='requests', sub_user_journey='household-code'
         )
-        self.get_requestcode_address_confirmation_hh_cy = self.app.router['RequestCodeConfirmAddress:get'].url_for(
-            request_type='household', display_region='cy'
+        self.post_request_code_confirm_address_hh_cy = self.app.router['CommonConfirmAddress:post'].url_for(
+            display_region='cy', user_journey='requests', sub_user_journey='household-code'
         )
-        self.get_requestcode_address_confirmation_hh_ni = self.app.router['RequestCodeConfirmAddress:get'].url_for(
-            request_type='household', display_region='ni'
-        )
-        self.post_requestcode_address_confirmation_hh_en = self.app.router['RequestCodeConfirmAddress:post'].url_for(
-            request_type='household', display_region='en'
-        )
-        self.post_requestcode_address_confirmation_hh_cy = self.app.router['RequestCodeConfirmAddress:post'].url_for(
-            request_type='household', display_region='cy'
-        )
-        self.post_requestcode_address_confirmation_hh_ni = self.app.router['RequestCodeConfirmAddress:post'].url_for(
-            request_type='household', display_region='ni'
+        self.post_request_code_confirm_address_hh_ni = self.app.router['CommonConfirmAddress:post'].url_for(
+            display_region='ni', user_journey='requests', sub_user_journey='household-code'
         )
 
         self.get_requestcode_individual_en = self.app.router['RequestCode:get'].url_for(
@@ -712,23 +704,14 @@ class RHTestCase(AioHTTPTestCase):
             display_region='ni', user_journey='requests', sub_user_journey='individual-code'
         )
 
-        self.get_requestcode_address_confirmation_hi_en = self.app.router['RequestCodeConfirmAddress:get'].url_for(
-            request_type='individual', display_region='en'
+        self.post_request_code_confirm_address_hi_en = self.app.router['CommonConfirmAddress:post'].url_for(
+            display_region='en', user_journey='requests', sub_user_journey='individual-code'
         )
-        self.get_requestcode_address_confirmation_hi_cy = self.app.router['RequestCodeConfirmAddress:get'].url_for(
-            request_type='individual', display_region='cy'
+        self.post_request_code_confirm_address_hi_cy = self.app.router['CommonConfirmAddress:post'].url_for(
+            display_region='cy', user_journey='requests', sub_user_journey='individual-code'
         )
-        self.get_requestcode_address_confirmation_hi_ni = self.app.router['RequestCodeConfirmAddress:get'].url_for(
-            request_type='individual', display_region='ni'
-        )
-        self.post_requestcode_address_confirmation_hi_en = self.app.router['RequestCodeConfirmAddress:post'].url_for(
-            request_type='individual', display_region='en'
-        )
-        self.post_requestcode_address_confirmation_hi_cy = self.app.router['RequestCodeConfirmAddress:post'].url_for(
-            request_type='individual', display_region='cy'
-        )
-        self.post_requestcode_address_confirmation_hi_ni = self.app.router['RequestCodeConfirmAddress:post'].url_for(
-            request_type='individual', display_region='ni'
+        self.post_request_code_confirm_address_hi_ni = self.app.router['CommonConfirmAddress:post'].url_for(
+            display_region='ni', user_journey='requests', sub_user_journey='individual-code'
         )
 
         self.get_requestcode_entermobile_hh_en = self.app.router['RequestCodeEnterMobile:get'].url_for(
@@ -981,19 +964,22 @@ class RHTestCase(AioHTTPTestCase):
 
         self.post_start_unlinked_select_address_en = \
             self.app.router['CommonSelectAddress:post'].url_for(display_region='en', user_journey='start',
-                                                                       sub_user_journey='unlinked')
+                                                                sub_user_journey='unlinked')
         self.post_start_unlinked_select_address_cy = \
             self.app.router['CommonSelectAddress:post'].url_for(display_region='cy', user_journey='start',
-                                                                       sub_user_journey='unlinked')
+                                                                sub_user_journey='unlinked')
         self.post_start_unlinked_select_address_ni = \
             self.app.router['CommonSelectAddress:post'].url_for(display_region='ni', user_journey='start',
-                                                                       sub_user_journey='unlinked')
+                                                                sub_user_journey='unlinked')
         self.post_start_unlinked_confirm_address_en = \
-            self.app.router['StartUnlinkedConfirmAddress:post'].url_for(display_region='en')
+            self.app.router['CommonConfirmAddress:post'].url_for(display_region='en', user_journey='start',
+                                                                 sub_user_journey='unlinked')
         self.post_start_unlinked_confirm_address_cy = \
-            self.app.router['StartUnlinkedConfirmAddress:post'].url_for(display_region='cy')
+            self.app.router['CommonConfirmAddress:post'].url_for(display_region='cy', user_journey='start',
+                                                                 sub_user_journey='unlinked')
         self.post_start_unlinked_confirm_address_ni = \
-            self.app.router['StartUnlinkedConfirmAddress:post'].url_for(display_region='ni')
+            self.app.router['CommonConfirmAddress:post'].url_for(display_region='ni', user_journey='start',
+                                                                 sub_user_journey='unlinked')
         self.post_start_unlinked_address_is_linked_en = \
             self.app.router['StartAddressHasBeenLinked:post'].url_for(display_region='en')
         self.post_start_unlinked_address_is_linked_cy = \
